backend/requirements.txt
django-cors-headers==3.5.0
django==3.1.3
psycopg2-binary==2.8.6
djangorestframework==3.12.2